refactor(core): move Context class from core/context.ts to top-level context.ts and update import paths
- Move context.ts from packages/core/src/core/ to packages/core/src/
- Update import path in constants.ts from './context' to '../context'
