+++ packages/core/src/core/constants.ts
@@ -1,6 +1,6 @@
 import type {
     Context,
-} from './context';
+} from '../context';
 
 import {
     interpolateRotation,
