+++ packages/core/src/core/factory.ts
@@ -1,12 +1,9 @@
 import type {
     BaseState,
-    MeasureTextOptions,
-} from './context';
-
-import type {
     Context,
     ContextOptions,
-} from './context';
+    MeasureTextOptions,
+} from '../context';
 
 /** Platform-specific function implementations injected at runtime. */
 export interface FactoryOptions {
