+++ packages/core/src/elements/text.ts
@@ -1,13 +1,19 @@
 import {
-    Context,
     Element,
-    measureText,
 } from '../core';
 
 import type {
     BaseElementState,
     ElementOptions,
 } from '../core';
+
+import {
+    measureText,
+} from '../context';
+
+import type {
+    Context,
+} from '../context';
 
 import {
     Box,
